// ui/src/generated/PresetDevice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BuiltinDeviceType } from "./BuiltinDeviceType";
import type { PluginFormat } from "./PluginFormat";

/**
 * The device type a preset is for.
 */
export type PresetDevice = { "type": "Builtin", device: BuiltinDeviceType, } | { "type": "Plugin", format: PluginFormat, plugin_id: string, 
/**
 * Display only.
 */
name: string, vendor: string, };